ref: separate main package in example folder
