+++ makefile
@@ -25,4 +25,4 @@ clean:
 	rm -rf $(GOGEN_OUTPUT) && rm -rf $(JAVRO_OUTPUT)
 
 parser:
-	go run examples/bundle_hl7_oru/main.go
+	go run examples/bundle_hl7_oru/*.go
